recommender-ai-service/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
requests==2.32.3
python-dotenv==1.0.1
